requirements.txt
attrs==23.1.0
Babel==2.9.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
exceptiongroup==1.1.3
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.11
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pluggy==1.3.0
pyTelegramBotAPI==4.12.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
